requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
dominate==2.3.1
eventlet==0.23.0
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-MQTT==1.0.4
Flask-SocketIO==3.0.1
greenlet==0.4.14
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
netifaces==0.10.7
paho-mqtt==1.3.1
python-engineio==2.2.0
python-socketio==2.0.0
pywemo==0.4.28
requests==2.19.1
six==1.11.0
typing==3.6.4
urllib3==1.23
visitor==0.1.3
Werkzeug==0.14.1
